[Sparse]: Refactor algorithm layer again.
