+++ python/sglang/srt/mem_cache/sparsity/__init__.py
@@ -1,8 +1,8 @@
 from sglang.srt.mem_cache.sparsity.algorithms import (
     BaseSparseAlgorithm,
+    BaseSparseAlgorithmImpl,
     DeepSeekNSAAlgorithm,
     KnormPageAlgorithm,
-    SparseMode,
 )
 from sglang.srt.mem_cache.sparsity.backend import BackendAdaptor, FlashAttentionAdaptor
 from sglang.srt.mem_cache.sparsity.core import SparseConfig, SparseCoordinator
@@ -14,7 +14,7 @@ from sglang.srt.mem_cache.sparsity.factory import (
 
 __all__ = [
     "BaseSparseAlgorithm",
-    "SparseMode",
+    "BaseSparseAlgorithmImpl",
     "KnormPageAlgorithm",
     "DeepSeekNSAAlgorithm",
     "BackendAdaptor",
