+++ python/sglang/srt/mem_cache/sparsity/algorithms/__init__.py
@@ -1,17 +1,13 @@
 from sglang.srt.mem_cache.sparsity.algorithms.base_algorithm import (
     BaseSparseAlgorithm,
-    SparseMode,
+    BaseSparseAlgorithmImpl,
 )
 from sglang.srt.mem_cache.sparsity.algorithms.deepseek_nsa import DeepSeekNSAAlgorithm
-from sglang.srt.mem_cache.sparsity.algorithms.page_wise_algorithm import (
-    BasePageWiseAlgorithm,
-    KnormPageAlgorithm,
-)
+from sglang.srt.mem_cache.sparsity.algorithms.knorm_algorithm import KnormPageAlgorithm
 
 __all__ = [
     "BaseSparseAlgorithm",
-    "SparseMode",
-    "BasePageWiseAlgorithm",
+    "BaseSparseAlgorithmImpl",
     "KnormPageAlgorithm",
     "DeepSeekNSAAlgorithm",
 ]
